mode fifo
git-svn-id: http://svn.ives.fr/svn-libs-dev/asterisk/fontventa/trunk@8036 4a32156a-8e32-406b-b38d-595c535062e7

+++ libmedikit/framebuffer.cpp
@@ -170,7 +170,7 @@ struct ast_frame * AstFrameBuffer::Wait()
 	return rtp;
 }
 
-struct AstFb *AstFbCreate(DWORD maxWaitTime, int blocking, int fifo)
+struct AstFb *AstFbCreate(unsigned long maxWaitTime, int blocking, int fifo)
 {
 	AstFrameBuffer * fb = new AstFrameBuffer((bool) blocking, (bool) fifo);
 	if (fb)
